fix: add ternary to show detail or edit
