+++ src/components/Detail/ViewDetail.js
@@ -7,7 +7,7 @@ import InputLabel from '@material-ui/core/InputLabel';
 class ViewDetail extends Component {
   render() {
     const bookId = this.props.match.params.id;
-    const { editBook, goBack, bookList } = this.props;
+    const { goBack, bookList } = this.props;
 
     if (bookList.length > 0 && bookId <= bookList.length) {
       const myBook = bookList[bookId - 1];
@@ -16,34 +16,29 @@ class ViewDetail extends Component {
       return (
         <Fragment>
           <div className="detail__wrapper">
-            <div className="btn__container">
-              <button className="form__btn" onClick={this.props.changeMe} >EDITAME</button>
-              <input type="submit" value="Enviar" className="form__btn" />
-              <button className="form__btn form__btn--close" onClick={goBack} >Cerrar</button>
-            </div>
             <ul className="list__view-details">
               <li className="list__item">
                 <InputLabel id="list__item--title" htmlFor="standard-read-only-input">Título</InputLabel>
                 <TextField className="list__item--content" id="standard-read-only-input" defaultValue={title} margin="normal" InputProps={{
-                  readOnly: Boolean(editBook),
+                  readOnly: true,
                 }} />
               </li>
               <li className="list__item">
                 <InputLabel id="list__item--title" htmlFor="standard-read-only-input">Autor</InputLabel>
                 <TextField className="list__item--content" id="standard-read-only-input" defaultValue={author} margin="normal" InputProps={{
-                readOnly: Boolean(editBook),
+                readOnly: true,
                 }} />
               </li>
               <li className="list__item">
                 <InputLabel id="list__item--title" htmlFor="standard-read-only-input">ISBN</InputLabel>
                 <TextField className="list__item--content" id="standard-read-only-input" defaultValue={ISBN} margin="normal" InputProps={{
-                  readOnly: Boolean(editBook),
+                  readOnly: true,
                 }} />
               </li>
               <li className="list__item">
                 <InputLabel id="list__item--title" htmlFor="standard-read-only-input">Tipo</InputLabel>
                 <TextField className="list__item--content" id="standard-read-only-input" defaultValue={type} margin="normal" InputProps={{
-                  readOnly: Boolean(editBook),
+                  readOnly: true,
                 }} />
               </li>
               <li className="list__item">
@@ -61,7 +56,7 @@ class ViewDetail extends Component {
               <li className="list__item">
                 <InputLabel id="list__item--title" htmlFor="standard-read-only-input">Estado</InputLabel>
                 <TextField className="list__item--content" id="standard-read-only-input" defaultValue={status} margin="normal" InputProps={{
-                  readOnly: Boolean(editBook),
+                  readOnly: true,
                 }} />
               </li>
             </ul>
